refactor; Moved out difference module
