+++ cognite/powerops/resync/diff/__init__.py
@@ -1,0 +1,17 @@
+"""
+This package is used to compare instances of models. It is used to determine if a model has changed, and if so, what
+has changed.
+"""
+
+from .data_classes import FieldIds, FieldDifference, ModelDifference, ModelDifferences, FieldSummary, Change
+from .core import model_difference
+
+__all__ = [
+    "FieldIds",
+    "FieldDifference",
+    "ModelDifference",
+    "ModelDifferences",
+    "FieldSummary",
+    "Change",
+    "model_difference",
+]
