+++ cognite/powerops/resync/diff/data_classes.py
@@ -216,7 +216,11 @@ class ModelDifferences:
             change_count = sum(len(change.changed) for change in changes)
             report.append(f"* **{name}**: {added} added, {removed} removed, {change_count} changed")
         list_ = "\n".join(report)
-        return f"""{'**' if no_headers else '##'} Summary {'**' if no_headers else ''}
+        if no_headers:
+            top_row = "**Summary**"
+        else:
+            top_row = "## Summary"
+        return f"""{top_row}
 {list_ if list_ else 'No changes'}
 """
