+++ cognite/powerops/resync/models/base/__init__.py
@@ -9,6 +9,8 @@ from .data_model import DataModel
 from .model import Model, T_Model
 from .cdf_resources import CDFSequence, CDFFile, CDFResource
 from .graph_ql import PowerOpsGraphQLModel
+from .resource_type import ResourceType
+
 
 __all__ = [
     "AssetModel",
@@ -22,4 +24,5 @@ __all__ = [
     "CDFFile",
     "CDFResource",
     "PowerOpsGraphQLModel",
+    "ResourceType",
 ]
