--- cognite/powerops/resync/models/diff/__init__.py
@@ -1,9 +1,0 @@
-"""
-This package is used to compare instances of models. It is used to determine if a model has changed, and if so, what
-has changed.
-"""
-
-from ._changes import FieldIds, FieldDifference, ModelDifference, ModelDifferences, FieldSummary, Change
-
-
-__all__ = ["FieldIds", "FieldDifference", "ModelDifference", "ModelDifferences", "FieldSummary", "Change"]
